# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shijie Learning Platform API",
    "description": "课堂 → 练习 → 巩固 智能学习闭环平台后端。Contract-first; /openapi.json 是唯一机器契约。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/practice-sets": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Practice Set Create",
        "operationId": "practice_set_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeSetCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice-sets/{set_id}": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Practice Set Get",
        "operationId": "practice_set_get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice-sets/{set_id}/attempts": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Attempt Create",
        "operationId": "attempt_create",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attempts": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Attempt Standalone",
        "description": "Review-flow attempts (no practice set).",
        "operationId": "attempt_create_standalone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttemptIn": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "is_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Correct"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "hint_count": {
            "type": "integer",
            "title": "Hint Count",
            "default": 0
          },
          "answer_change_count": {
            "type": "integer",
            "title": "Answer Change Count",
            "default": 0
          },
          "grading_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grading Source"
          },
          "review_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Task Id"
          }
        },
        "type": "object",
        "required": [
          "question_id"
        ],
        "title": "AttemptIn"
      },
      "AttemptOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "practice_set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice Set Id"
          },
          "is_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Correct"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "attempted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Attempted At"
          },
          "created_memory_state": {
            "type": "boolean",
            "title": "Created Memory State",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "practice_set_id",
          "is_correct",
          "score",
          "attempted_at"
        ],
        "title": "AttemptOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Option": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "key",
          "text"
        ],
        "title": "Option"
      },
      "PracticeSetCreateIn": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "random"
          },
          "count": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Count",
            "default": 5
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "knowledge_point_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledge Point Ids",
            "default": []
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "wrong_only": {
            "type": "boolean",
            "title": "Wrong Only",
            "default": false
          },
          "favorite_only": {
            "type": "boolean",
            "title": "Favorite Only",
            "default": false
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "title": "PracticeSetCreateIn"
      },
      "PracticeSetItemOut": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionDetailOut"
          }
        },
        "type": "object",
        "required": [
          "position",
          "question"
        ],
        "title": "PracticeSetItemOut"
      },
      "PracticeSetOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PracticeSetItemOut"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "mode",
          "filters",
          "created_at"
        ],
        "title": "PracticeSetOut"
      },
      "QuestionDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type"
          },
          "stem": {
            "type": "string",
            "title": "Stem"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Option"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "solution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solution"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_favorite": {
            "type": "boolean",
            "title": "Is Favorite",
            "default": false
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "analysis_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Summary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject",
          "grade",
          "question_type",
          "stem",
          "answer",
          "solution",
          "difficulty",
          "content_hash",
          "created_at"
        ],
        "title": "QuestionDetailOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}